#164 renamed StrictEvaluationStrategy, added ExtendedEvaluationStrategy
