+++ core/http/workbench/src/main/webapp/transformations/create-memory.xsl
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE xsl:stylesheet>
 <xsl:stylesheet version="1.0"
-	xmlns:xsl="http://www.w3.org/1999/XSL/Transform"
-	xmlns:sparql="http://www.w3.org/2005/sparql-results#" xmlns="http://www.w3.org/1999/xhtml">
+	xmlns:xsl="http://www.w3.org/1999/XSL/Transform" xmlns:sparql="http://www.w3.org/2005/sparql-results#"
+	xmlns="http://www.w3.org/1999/xhtml">
 
 	<xsl:include href="../locale/messages.xsl" />
 
@@ -54,11 +54,10 @@
 							<xsl:value-of select="$repository-persist.label" />
 						</th>
 						<td>
-							<input type="radio" name="Persist" size="48"
-								value="true" checked="true" />
+							<input type="radio" name="Persist" size="48" value="true"
+								checked="true" />
 							<xsl:value-of select="$true.label" />
-							<input type="radio" name="Persist" size="48"
-								value="false" />
+							<input type="radio" name="Persist" size="48" value="false" />
 							<xsl:value-of select="$false.label" />
 						</td>
 						<td></td>
@@ -79,8 +78,13 @@
 						</th>
 						<td>
 							<select id="evalStratFactory" name="EvaluationStrategyFactory">
-								<option value="org.eclipse.rdf4j.query.algebra.evaluation.impl.SimpleEvaluationStrategyFactory">
-								Strict
+								<option selected="selected"
+									value="org.eclipse.rdf4j.query.algebra.evaluation.impl.StrictEvaluationStrategyFactory">
+									Strict
+								</option>
+								<option
+									value="org.eclipse.rdf4j.query.algebra.evaluation.impl.ExtendedEvaluationStrategyFactory">
+									Extended
 								</option>
 							</select>
 						</td>
@@ -91,7 +95,7 @@
 						<td>
 							<input type="button" value="{$cancel.label}" style="float:right"
 								data-href="repositories"
-                                onclick="document.location.href=this.getAttribute('data-href')" />
+								onclick="document.location.href=this.getAttribute('data-href')" />
 							<input id="create" type="button" value="{$create.label}"
 								onclick="checkOverwrite()" />
 						</td>
