+++ core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/QueryOptimizer.java
@@ -10,10 +10,10 @@ package org.eclipse.rdf4j.query.algebra.evaluation;
 import org.eclipse.rdf4j.query.BindingSet;
 import org.eclipse.rdf4j.query.Dataset;
 import org.eclipse.rdf4j.query.algebra.TupleExpr;
-import org.eclipse.rdf4j.query.algebra.evaluation.impl.SimpleEvaluationStrategy;
+import org.eclipse.rdf4j.query.algebra.evaluation.impl.StrictEvaluationStrategy;
 
 /**
- * Interface used by {@link SimpleEvaluationStrategy} prior to evalutating the query.
+ * Interface used by {@link StrictEvaluationStrategy} prior to evalutating the query.
  * 
  * @author James Leigh
  * @author Arjohn Kampman
