+++ core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/impl/EvaluationStrategyImpl.java
@@ -12,11 +12,11 @@ import org.eclipse.rdf4j.query.algebra.evaluation.TripleSource;
 import org.eclipse.rdf4j.query.algebra.evaluation.federation.FederatedServiceResolver;
 
 /**
- * @deprecated since 4.0. Use {@link SimpleEvaluationStrategy} instead.
+ * @deprecated since 4.0. Use {@link StrictEvaluationStrategy} instead.
  * @author Jeen Broekstra
  */
 @Deprecated
-public class EvaluationStrategyImpl extends SimpleEvaluationStrategy {
+public class EvaluationStrategyImpl extends StrictEvaluationStrategy {
 
 	public EvaluationStrategyImpl(TripleSource tripleSource, Dataset dataset,
 			FederatedServiceResolver serviceResolver)
