+++ core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/impl/StrictEvaluationStrategyFactory.java
@@ -1,3 +1,10 @@
+/*******************************************************************************
+ * Copyright (c) 2016 Eclipse RDF4J contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Distribution License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *******************************************************************************/
 package org.eclipse.rdf4j.query.algebra.evaluation.impl;
 
 import org.eclipse.rdf4j.query.Dataset;
@@ -7,13 +14,16 @@ import org.eclipse.rdf4j.query.algebra.evaluation.TripleSource;
 import org.eclipse.rdf4j.query.algebra.evaluation.federation.FederatedServiceResolver;
 import org.eclipse.rdf4j.query.algebra.evaluation.federation.FederatedServiceResolverClient;
 
-public class SimpleEvaluationStrategyFactory implements EvaluationStrategyFactory, FederatedServiceResolverClient {
+public class StrictEvaluationStrategyFactory
+		implements EvaluationStrategyFactory, FederatedServiceResolverClient
+{
+
 	private FederatedServiceResolver serviceResolver;
 
-	public SimpleEvaluationStrategyFactory() {
+	public StrictEvaluationStrategyFactory() {
 	}
 
-	public SimpleEvaluationStrategyFactory(FederatedServiceResolver resolver) {
+	public StrictEvaluationStrategyFactory(FederatedServiceResolver resolver) {
 		this.serviceResolver = resolver;
 	}
 
@@ -28,6 +38,6 @@ public class SimpleEvaluationStrategyFactory implements EvaluationStrategyFactor
 
 	@Override
 	public EvaluationStrategy createEvaluationStrategy(Dataset dataset, TripleSource tripleSource) {
-		return new SimpleEvaluationStrategy(tripleSource, dataset, serviceResolver);
+		return new StrictEvaluationStrategy(tripleSource, dataset, serviceResolver);
 	}
 }
