+++ core/queryalgebra/evaluation/src/test/java/org/eclipse/rdf4j/query/algebra/evaluation/impl/SimpleEvaluationStrategyTest.java
@@ -31,7 +31,7 @@ public class SimpleEvaluationStrategyTest {
 	public void setUp()
 		throws Exception
 	{
-		strategy = new SimpleEvaluationStrategy(new EmptyTripleSource(), null);
+		strategy = new StrictEvaluationStrategy(new EmptyTripleSource(), null);
 	}
 
 	/**
